+++ inst/paramtest/test_paramtest_classif.ranger.R
@@ -35,8 +35,8 @@ test_that("predict classif.ranger", {
   learner = lrn("classif.ranger")
   fun = ranger:::predict.ranger
   exclude = c(
-    "quantiles", # required type not supported in mlr3
-    "what", # required type (quantiles) not supported in mlr3
+    "quantiles", # not supported by classification
+    "what", # not supported by classification
     "predict.all", # not supported in mlr3
     "formula", # handled via mlr3
     "object", # handled via mlr3