+++ inst/paramtest/test_paramtest_regr.ranger.R
@@ -18,7 +18,8 @@ test_that("regr.ranger", {
     "status.variable.name", # handled via mlr3
     "classification", # handled via mlr3
     "mtry.ratio", # custom hyperpar
-    "time.interest" # survival only
+    "time.interest", # survival only
+    "quantreg" # handled by predict_type
   )
 
   ParamTest = run_paramtest(learner, fun, exclude, tag = "train")
@@ -34,8 +35,8 @@ test_that("predict regr.ranger", {
   learner = lrn("regr.ranger")
   fun = ranger:::predict.ranger
   exclude = c(
-    "quantiles", # required type not supported in mlr3
-    "what", # required type (quantiles) not supported in mlr3
+    "quantiles", # handled via mlr3
+    "what", # not supported in mlr3
     "predict.all", # not supported in mlr3
     "formula", # handled via mlr3
     "object", # handled via mlr3